+++ apps/studio/CHANGELOG.md
@@ -1,5 +1,27 @@
 # vite-studio
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [95a15b3]
+- Updated dependencies [9e62382]
+  - @tinloof/sanity-studio@1.3.1
+
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [2a263d7]
+- Updated dependencies [a09d24f]
+- Updated dependencies [29a848a]
+- Updated dependencies [4289934]
+- Updated dependencies [4c92cc8]
+- Updated dependencies [ac331cf]
+- Updated dependencies [7dd9046]
+- Updated dependencies [ab53988]
+  - @tinloof/sanity-studio@1.3.0
+
 ## 1.0.10
 
 ### Patch Changes