+++ examples/hello-world/CHANGELOG.md
@@ -4,6 +4,28 @@
 
 ### Patch Changes
 
+- Updated dependencies [95a15b3]
+- Updated dependencies [9e62382]
+  - @tinloof/sanity-studio@1.3.1
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [2a263d7]
+- Updated dependencies [a09d24f]
+- Updated dependencies [29a848a]
+- Updated dependencies [4289934]
+- Updated dependencies [4c92cc8]
+- Updated dependencies [ac331cf]
+- Updated dependencies [7dd9046]
+- Updated dependencies [ab53988]
+  - @tinloof/sanity-studio@1.3.0
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [a3677bb]
   - @tinloof/sanity-studio@1.2.1